feat: implement order placement flow and add request validation
* Implemented placeOrder logic with Order and OrderItem creation
* Added order amount calculation from cart items
* Cleared cart items after successful order placement
* Added DTO validation using jakarta validation annotations
* Added global validation exception handling

+++ src/main/java/com/mamoon/project/food_ordering_app/dto/CartItemResponseDTO.java
@@ -12,7 +12,7 @@ public class CartItemResponseDTO {
 
     Long foodItemId;
     String foodItemName;
-    Double foodItemPrice;
-   Long quantity;
-    Double subtotal;
+    BigDecimal foodItemPrice;
+    Long quantity;
+    BigDecimal subtotal;
 }
